// nodedoc/handle/nestjs/case/nest-demo/src/enum/config.enum.ts
export enum ConfigEnum {
  HOST = 'HOST',
  PORT = 'PORT',
  DB_DATABASE = 'DB_DATABASE',
  DB_ACCOUNT = 'DB_ACCOUNT',
  DB_PASSWD = 'DB_PASSWD',
  DB_HOST = 'DB_HOST',
  DB_PORT = 'DB_PORT',
  DB_SYNC = 'DB_SYNC',
  DB_TYPE = 'DB_TYPE',

  REDIS_URL = 'REDIS_URL',
  REDIS_PASSWD = 'REDIS_PASSWD',
  AUTH_SECRET_KEY = 'AUTH_SECRET_KEY',
  SERVER_HOST = 'SERVER_HOST',
}